+++ tests/testthat/test_recursive/logic/test-logic_sample.R
@@ -1,0 +1,7 @@
+test_that("multiplication works", {
+  expect_equal(2 * 2, 4)
+})
+
+test_that("a failed test example", {
+  expect_true(FALSE)
+})