+++ NEWS.md
@@ -1,5 +1,7 @@
 # rhino (development version)
 
+1. `test_r()` support for unit tests with a folder/directory structure.
+
 # [rhino 1.3.0](https://github.com/Appsilon/rhino/releases/tag/v1.3.0)
 
 1. Rhino now works with `shinytest2` out of the box.